simplify: one owner per fact, after the three-reviewer pass
The playback engine already knows the buffer of the over in progress, so
the server no longer keeps a second copy in step with it; asking is
cheaper than agreeing. The six identical little-endian writes in the ping
builders became one helper. The wire format the Python tools share -
frame composition, the 7-bit wait codec, the streaming parser - moved
into cwnet_wire.py, which both the echo and the test client import: two
copies of an idea of the wire drift, and a tool that has drifted lies
about the protocol instead of testing it.

clock_wire_ms() keeps its place but loses a comment that promised a use
the architecture forbids: the core masks its own timestamps because it
cannot depend on host/platform. It is there for the Windows client of
#68, and now says so.

keyer_cwnet's module brief now knows about the server core, the playback
engine and the parser's skipped state.

Declined: extracting a generic ring buffer shared by the client's receive
FIFO and the engine's. It is six lines of index arithmetic; in C the
generic version is macros, reads worse, and would edit firmware code
pinned against the capture. The remedy costs more than the duplication.

294 tests green in both variants, the host loopback green, and the
reference over still plays at 48, 48 and 144 ms with the tail at 100.
